pyproject.toml
[project]
name = "eyetest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.8",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "pyyaml>=6.0.3",
    "scikit-image>=0.26.0",
    "scikit-learn>=1.8.0",
    "torch>=2.8.0",
]
